requirements.txt
gunicorn
flask
requests
flask-cors
fastapi
uvicorn
pydantic
pydantic-settings
pytest
supabase
python-dotenv
eth-utils
web3
